[worker] wip with webcrypto api
