+++ packages/worker-api/src/encointerWorker.ts
@@ -1,9 +1,6 @@
 import type {u32, u64, Vec} from '@polkadot/types';
 import {communityIdentifierFromString} from '@encointer/util';
 
-// @ts-ignore
-import NodeRSA from '@learntheropes/node-rsa';
-
 import type {
     CommunityIdentifier,
     MeetupIndexType,
